Task 7/backend/requirements.txt
Django==4.2.2
djangorestframework
pyyaml
requests
asgiref==3.6.0
dj-database-url==2.0.0
django-cors-headers==4.0.0
gunicorn==20.1.0
psycopg==3.1.8
psycopg2==2.9.5
sqlparse==0.4.2
dj-database-url